display, inventory, sprites, dialogue
I added a few sprites including the rocket ship, lab, and government building, also the overworld map looks better now. I'm working on a few systems which are currently unfinished including the dialogue, inventory, and display settings, I spent a lot of time trying to figure out why the game was so tiny. Still messing with the positionality of items on the screen; they shouldn't follow the camera around...