LADDER!! and initialization screen
i ttook me thee hours, but I got a ladder to fnction in the game this includes fixes to wandering inventory sprites and item use functionality. I created an initialization screen with a game studio logo, but it's jacked up, and the menu screen has bad dimensions again. The mechanic character is now interactable, but the dialogue and text box are out of alignment