pause screen start and repository select
the pause screen is super janky and only on the lab room for now, but it's a start. i'll work on adding the right buttons and proportions on it next. I used some tutorial code and that's why it doesn't match. I mapped it to P and ESC. I also drew a repository room for the player to warp into from the telescope. The planets are speak_block objects. I wrote dummy text for the planet properties as well