fixed pause screen and score/lives counter
there isn't functionality for the score/lives counter yet, but they are in the right place in all of the rooms! also working on a paused screen that currently only operates in the lab room. i tried to reuse the menu screen buttons but it got too tricky becuase the pause screen isn't a room and therefore I can't put the button objects in it. A lot of troubleshooting and trying to make things work today means that there's a lot of extra things in the project file that don't need to be there...oops