lab and gov sprites, plus pi paper problems
i did sprites for the stuff in the lab and government rooms, however I'm having trouble with making the pi paper dissapear once the player clicks on the item. the same funciton will have to be used for the blueprint. also the rocket body is pick-up-able now but doesn't dissapear at the right time...