level 3 is good enough
i figured out how to add loop functionality for the quiz question. also hid the ladder and the rocket wing in the trees. there are only three pieces of information for level 3 now. I also put walls on teh repository planets, however the sprite for the engine isn't visible so I'll have to add something environmental for that, or maybe put it in the overworld after the right planet is selected...